+++ frontend/src/pages/MemberTemplatePage/components/CategoryEditModal/CategoryEditModal.tsx
@@ -1,15 +1,12 @@
-import { css } from '@emotion/react';
-import { PropsWithChildren, useEffect, useRef, useState } from 'react';
+import { useState } from 'react';
 
-import { PencilIcon, SpinArrowIcon, TrashcanIcon, DragIcon } from '@/assets/images';
-import { Text, Modal, Input, Flex, Button } from '@/components';
+import { Text, Modal, Flex, Button } from '@/components';
 import { useCategoryNameValidation } from '@/hooks/category';
 import { useCategoryEditMutation } from '@/queries/categories';
 import { validateCategoryName } from '@/service/validates';
-import { ICON_SIZE } from '@/style/styleConstants';
-import { theme } from '@/style/theme';
 import type { Category } from '@/types';
 
+import CategoryItems from './CategoryItems';
 import * as S from './CategoryEditModal.style';
 
 interface CategoryEditModalProps {
@@ -54,14 +51,25 @@ const CategoryEditModal = ({
 
     if (isNewCategory(id)) {
       setNewCategories((prev) => prev.map((category) => (category.id === id ? { ...category, name } : category)));
-    } else {
-      const ordinal = editedCategories.find((category) => category.id === id)?.ordinal as number;
 
-      setEditedCategories((prev) => [...prev, { id, name, ordinal }]);
+      return;
     }
+
+    //TODO: targetCategory 네이밍 바꾸기
+    const targetCategory = editedCategories.find((category) => category.id === id);
+
+    if (targetCategory) {
+      setEditedCategories((prev) => prev.map((category) => (category.id === id ? { ...category, name } : category)));
+
+      return;
+    }
+
+    const ordinal = categories.find((category) => category.id === id)?.ordinal as number;
+
+    setEditedCategories((prev) => [...prev, { id, name, ordinal }]);
   };
 
-  const handleDrag = (categories: Category[]) => {
+  const handleOrdinalChange = (categories: Category[]) => {
     const updatedCategories: Category[] = [];
     const updatedNewCategories: Category[] = [];
 
@@ -114,7 +122,7 @@ const CategoryEditModal = ({
         ? categories[categories.length - 1].id + newCategories.length + 1
         : newCategories.length + 1;
 
-    const ordinal = categories.length + newCategories.length;
+    const ordinal = categories.length + 1 + newCategories.length;
 
     setNewCategories((prev) => [...prev, { id, name: '', ordinal }]);
     setEditingCategoryId(id);
@@ -154,7 +162,7 @@ const CategoryEditModal = ({
           <CategoryItems
             categories={categories}
             newCategories={newCategories}
-            handleDrag={handleDrag}
+            handleDrag={handleOrdinalChange}
             editedCategories={editedCategories}
             categoriesToDelete={deleteCategoryIds}
             editingCategoryId={editingCategoryId}
@@ -193,303 +201,5 @@ const CategoryEditModal = ({
     </Modal>
   );
 };
-
-interface CategoryItemsProps {
-  categories: Category[];
-  newCategories: Category[];
-  editedCategories: Category[];
-  categoriesToDelete: number[];
-  editingCategoryId: number | null;
-  invalidIds: number[];
-  isNewCategory: (id: number) => boolean;
-  handleDrag: (categories: Category[]) => void;
-  onEditClick: (id: number) => void;
-  onDeleteClick: (id: number) => void;
-  onRestoreClick: (id: number) => void;
-  onNameInputChange: (id: number, name: string) => void;
-  onNameInputBlur: (id: number) => void;
-}
-
-const CategoryItems = ({
-  categories,
-  newCategories,
-  editedCategories,
-  categoriesToDelete,
-  editingCategoryId,
-  invalidIds,
-  isNewCategory,
-  handleDrag,
-  onEditClick,
-  onDeleteClick,
-  onRestoreClick,
-  onNameInputChange,
-  onNameInputBlur,
-}: CategoryItemsProps) => {
-  const categoriesMap = new Map();
-
-  [...categories, ...editedCategories, ...newCategories].forEach((category) => {
-    categoriesMap.set(category.id, category);
-  });
-
-  const initOrderedCategoriesArray = Array.from(categoriesMap.values()).sort((a, b) => a.ordinal - b.ordinal);
-
-  const [orderedCategories, setOrderedCategories] = useState(initOrderedCategoriesArray);
-
-  useEffect(() => {
-    const categoriesMap = new Map();
-
-    [...categories, ...editedCategories, ...newCategories].forEach((category) => {
-      categoriesMap.set(category.id, category);
-    });
-
-    const orderedCategoriesArray = Array.from(categoriesMap.values()).sort((a, b) => a.ordinal - b.ordinal);
-
-    setOrderedCategories(orderedCategoriesArray);
-  }, [newCategories, editedCategories, categories]);
-
-  const dragItem = useRef<number | null>(null);
-  const dragOverItem = useRef<number | null>(null);
-
-  const handleDragStart = (e: React.DragEvent<HTMLDivElement>, position: number) => {
-    dragItem.current = position;
-    e.currentTarget.style.opacity = '0.5';
-  };
-
-  const handleDragEnter = (e: React.DragEvent<HTMLDivElement>, position: number) => {
-    dragOverItem.current = position;
-    e.currentTarget.style.backgroundColor = '#f5f5f5';
-  };
-
-  const handleDragEnd = (e: React.DragEvent<HTMLDivElement>) => {
-    if (dragItem.current === null || dragOverItem.current === null) {
-      return;
-    }
-
-    e.currentTarget.style.opacity = '1';
-    e.currentTarget.style.backgroundColor = '';
-
-    const copyListItems = [...orderedCategories];
-    const dragItemContent = copyListItems[dragItem.current];
-
-    copyListItems.splice(dragItem.current, 1);
-    copyListItems.splice(dragOverItem.current, 0, dragItemContent);
-
-    const updatedItems = copyListItems.map((item, index) => ({
-      ...item,
-      ordinal: index + 1,
-    }));
-
-    dragItem.current = null;
-    dragOverItem.current = null;
-    handleDrag(updatedItems);
-    setOrderedCategories(updatedItems);
-  };
-
-  const handleDragLeave = (e: React.DragEvent<HTMLDivElement>) => {
-    e.currentTarget.style.backgroundColor = '';
-  };
-
-  return (
-    <>
-      {orderedCategories.map(({ id, name }, index) => (
-        <S.EditCategoryItem
-          key={id}
-          hasError={invalidIds.includes(id)}
-          draggable
-          onDragStart={(e) => handleDragStart(e, index)}
-          onDragEnter={(e) => handleDragEnter(e, index)}
-          onDragEnd={handleDragEnd}
-          onDragLeave={handleDragLeave}
-          onDragOver={(e) => e.preventDefault()}
-        >
-          {isNewCategory(id) ? (
-            <NewCategoryItem
-              id={id}
-              name={name}
-              isEditing={editingCategoryId === id}
-              onChange={(e) => onNameInputChange(id, e.target.value)}
-              onBlur={() => onNameInputBlur(id)}
-              onKeyDown={(e) => {
-                if (e.key === 'Enter') {
-                  onNameInputBlur(id);
-                }
-              }}
-              onEditClick={() => onEditClick(id)}
-              onDeleteClick={() => onDeleteClick(id)}
-            />
-          ) : (
-            <ExistingCategoryItem
-              id={id}
-              name={editedCategories.find((category) => category.id === id)?.name ?? name}
-              isEditing={editingCategoryId === id}
-              isDeleted={categoriesToDelete.includes(id)}
-              onChange={(e) => onNameInputChange(id, e.target.value)}
-              onBlur={() => onNameInputBlur(id)}
-              onKeyDown={(e) => {
-                if (e.key === 'Enter') {
-                  onNameInputBlur(id);
-                }
-              }}
-              onEditClick={() => onEditClick(id)}
-              onDeleteClick={() => onDeleteClick(id)}
-              onRestoreClick={() => onRestoreClick(id)}
-            />
-          )}
-        </S.EditCategoryItem>
-      ))}
-    </>
-  );
-};
-
-interface ExistingCategoryItemProps {
-  id: number;
-  name: string;
-  isEditing: boolean;
-  isDeleted: boolean;
-  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  onBlur: () => void;
-  onKeyDown: (e: React.KeyboardEvent) => void;
-  onEditClick: (id: number) => void;
-  onDeleteClick: (id: number) => void;
-  onRestoreClick: (id: number) => void;
-}
-
-const ExistingCategoryItem = ({
-  id,
-  name,
-  isEditing,
-  isDeleted,
-  onChange,
-  onBlur,
-  onKeyDown,
-  onEditClick,
-  onDeleteClick,
-  onRestoreClick,
-}: ExistingCategoryItemProps) => (
-  <>
-    {isDeleted ? (
-      <>
-        <CategoryName>
-          <Text.Medium color={theme.color.light.analogous_primary_400} textDecoration='line-through'>
-            {name}
-          </Text.Medium>
-        </CategoryName>
-        <IconButtons restore onRestoreClick={() => onRestoreClick(id)} />
-      </>
-    ) : (
-      <>
-        <CategoryName>
-          {isEditing ? (
-            <CategoryNameInput value={name} onChange={onChange} onBlur={onBlur} onKeyDown={onKeyDown} />
-          ) : (
-            <Text.Medium color={theme.color.light.secondary_500} weight='bold'>
-              {name}
-            </Text.Medium>
-          )}
-        </CategoryName>
-        <IconButtons edit delete onEditClick={() => onEditClick(id)} onDeleteClick={() => onDeleteClick(id)} />
-      </>
-    )}
-  </>
-);
-
-interface NewCategoryItemProps {
-  id: number;
-  name: string;
-  isEditing: boolean;
-  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  onBlur: () => void;
-  onKeyDown: (e: React.KeyboardEvent) => void;
-  onEditClick: (id: number) => void;
-  onDeleteClick: (id: number) => void;
-}
-
-const NewCategoryItem = ({
-  id,
-  name,
-  isEditing,
-  onChange,
-  onBlur,
-  onKeyDown,
-  onEditClick,
-  onDeleteClick,
-}: NewCategoryItemProps) => (
-  <>
-    <CategoryName>
-      {isEditing ? (
-        <CategoryNameInput value={name} onChange={onChange} onBlur={onBlur} onKeyDown={onKeyDown} />
-      ) : (
-        <Text.Medium color={theme.color.light.secondary_500} weight='bold'>
-          {name}
-        </Text.Medium>
-      )}
-    </CategoryName>
-    <IconButtons edit delete onEditClick={() => onEditClick(id)} onDeleteClick={() => onDeleteClick(id)} />
-  </>
-);
-
-const CategoryName = ({ children }: PropsWithChildren) => (
-  <Flex align='center' width='100%' height='2.5rem'>
-    <DragIcon color={theme.color.light.secondary_400} css={{ marginRight: '0.5rem' }} />
-    {children}
-  </Flex>
-);
-
-interface CategoryNameInputProps {
-  value: string;
-  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  onBlur: () => void;
-  onKeyDown: (e: React.KeyboardEvent) => void;
-}
-
-const CategoryNameInput = ({ value, onChange, onBlur, onKeyDown }: CategoryNameInputProps) => (
-  <Input size='large' variant='text' style={{ width: '100%', height: '2.375rem' }}>
-    <Input.TextField
-      type='text'
-      value={value}
-      placeholder='카테고리 입력'
-      onChange={onChange}
-      onBlur={onBlur}
-      onKeyDown={onKeyDown}
-      autoFocus
-      css={css`
-        font-weight: bold;
-        &::placeholder {
-          font-weight: normal;
-          color: ${theme.color.light.secondary_400};
-        }
-      `}
-    />
-  </Input>
-);
-
-interface IconButtonsProps {
-  onRestoreClick?: () => void;
-  onEditClick?: () => void;
-  onDeleteClick?: () => void;
-  restore?: boolean;
-  edit?: boolean;
-  delete?: boolean;
-}
-
-const IconButtons = ({ onRestoreClick, onEditClick, onDeleteClick, restore, edit, delete: del }: IconButtonsProps) => (
-  <S.IconButtonContainer>
-    {restore && (
-      <S.IconButtonWrapper onClick={onRestoreClick}>
-        <SpinArrowIcon aria-label='카테고리 복구' />
-      </S.IconButtonWrapper>
-    )}
-    {edit && (
-      <S.IconButtonWrapper onClick={onEditClick}>
-        <PencilIcon width={ICON_SIZE.MEDIUM_LARGE} height={ICON_SIZE.MEDIUM_LARGE} aria-label='카테고리 이름 변경' />
-      </S.IconButtonWrapper>
-    )}
-    {del && (
-      <S.IconButtonWrapper onClick={onDeleteClick}>
-        <TrashcanIcon width={ICON_SIZE.MEDIUM_LARGE} height={ICON_SIZE.MEDIUM_LARGE} aria-label='카테고리 삭제' />
-      </S.IconButtonWrapper>
-    )}
-  </S.IconButtonContainer>
-);
 
 export default CategoryEditModal;
